overall: switch default synchronicity model to synchronous
Bring both sync and async models to the end user (test writer).

It work by keeping an event loop running the whole process and by
scheduling everything that is async to it, this way the main threa
can still perform blocking IO without bothering about `Bumble` IO &
gRPC server IO running in the background.

The default model choose to be sync one. For a test to correctly switch
to the async model, it shall decorate its function with the newly added
`@avatar.asynchronous` decorator. This decorator convert the function to
a synchronous one and force the devices in `pandora_devices` to use their
asynchronous gRPC channels instead of the sync ones.

See also:
 * Update `bt-test-interfaces` git submodule which include minor fixes
   to the host API and gRPC client API timeout handling.
 * Update `third-party/bumble` git submodule which include minor fixes
   to make the `host.WaitConnection` work.
 * Update of `example.py` to showcase the latest changes to the avatar
   test framework.
 * Fixed some data types in bumble host server.

Bug: 246217816
Test: python examples/example.py -c examples/simulated_bumble_bumble.yml

Change-Id: I2f058e747a1ec3f2219f9f410aa9b2d4fa2fb289