overall: rework bumble server loop
Rework the factory reset mechanism to only trigger gRPC server
close. The server loop then wait for the server termination and
start a fresh one. In the mean team, `ReadLocalAddress` in the
`setup_class` is actually waiting for this to be done without
being aware of it, since the server is not responding.

This particular mechanism make sure there is nothing more running
of the event loop. Before that change, we encountered zombie
timeouts poping from dead async generators.

Also introduce a new `parameterized` decorator to allow the write
of paramterized tests.

Change-Id: I4caf612b28081606f09f83fc870db89ba576e116