move reset from setup to teardown
The reset is currently being done in the setup_test
hence it is executed for each tests even if the test is skipped.
Doing the reset on the teardown let us skip the reset when it is
not desired such as after a skipped test.

This reduce considerably the overall time for the tests execution.